ejercicio_1/api/urls.py
from django.urls import include, path
from rest_framework import routers
from rest_framework.schemas import get_schema_view
from django.views.generic import TemplateView
from . import views

router = routers.DefaultRouter()
router.register(r'medidores', views.MedidorViewSet, basename='medidor', )
router.register(r'mediciones', views.MedicionViewSet,  basename='medicion')

# Wire up our API using automatic URL routing.
# Additionally, we include login URLs for the browsable API.
urlpatterns = [
    
    path('', include(router.urls)),
    path('api-auth/', include('rest_framework.urls', namespace='rest_framework')),
    path('openapi/', get_schema_view(
        title="Ejercicio_1 Service",
        description="API para Challenge Powermeter"
    ), name='openapi-schema'),
    path('docs/', TemplateView.as_view(
        template_name='documentation.html',
        extra_context={'schema_url':'openapi-schema'}
    ), name='swagger-ui'),
]